backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
pgvector
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
celery[redis]
python-multipart
httpx
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
sentry-sdk[fastapi]
slowapi
pytest
requests
ruff
google-generativeai
bcrypt==4.0.1
